define vertex & fragment shader strings in module files
